Fix the AvatarUpdate so that it updates in real time
